+++ lib/modules/forms/forms_service.dart
@@ -1,0 +1,9 @@
+﻿// Module: forms
+// Updated at: 2026-04-01 13:11:41
+// Commit: feat(forms): add bi-directional event bus handler for forms
+
+class formsComponentState {
+  final String commitRef = '11';
+  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
+  final DateTime timestamp = DateTime.parse('2026-04-01T13:11:41+05:00');
+}